cleanup - delete unnecessary reducer
