+++ src/constants/ActionTypes.js
@@ -1,4 +1,3 @@
-export const APPLY_DATA = 'APPLY_DATA'
 export const PROCESS_DATA_REQUEST = 'PROCESS_DATA_REQUEST'
 export const CHANGE_MAPPING = 'CHANGE_MAPPING'
 export const SELECTED_DIMENSIONVALUE = 'SELECTED_DIMENSIONVALUE'
